move some functions into utils
